requirements.txt
h5py==3.10.0
kornia==0.7.0
matplotlib==3.7.3
numpy==1.23.2
opencv_python==4.8.1.78
Pillow==9.0.1
Pillow==10.2.0
scikit_image==0.16.2
scikit_learn==1.3.2
scipy==1.11.4
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.2
